serveur-mc/requirements.txt
discord.py>=2.0.0
boto3
python-dotenv